requirements_picture.txt
fastapi
python-multipart
uvicorn
pyyaml

RPi.GPIO
gpiozero
numpy
pillow
spidev

waveshare-epaper
